requirements.txt
alembic==1.4.3
apache-airflow==1.10.14
apispec==3.3.2
argcomplete==1.12.2
attrs==20.3.0
Babel==2.9.0
cached-property==1.5.2
cattrs==1.0.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==6.7
colorama==0.4.4
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.36
cryptography==3.2.1
defusedxml==0.6.0
dill==0.3.3
dnspython==1.16.0
docutils==0.15.2
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==1.13.1
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-JWT-Extended==3.25.0
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.14
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
graphviz==0.14.2
gunicorn==20.0.4
idna==2.8
importlib-metadata==2.1.0
importlib-resources==1.5.0
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.11.2
json-merge-patch==0.2
jsonschema==3.2.0
kafka-python==2.0.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.18.0
natsort==7.1.0
numpy==1.18.5
packaging==20.7
pandas==0.25.3
pendulum==1.4.4
pep562==1.0
prison==0.1.0
psutil==5.7.3
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.4
pytzdata==2020.1
PyYAML==5.3.1
requests==2.23.0
setproctitle==1.1.10
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-JSONField==0.9.0
tabulate==0.8.7
tenacity==4.12.0
text-unidecode==1.3
thrift==0.13.0
typing==3.7.4.3
typing-extensions==3.7.4.3
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.11
Werkzeug==0.16.1
WTForms==2.3.3
zipp==1.2.0
zope.deprecation==4.4.0
